BDOG-1267: add in play 2.8
