+++ project/build.properties
@@ -14,4 +14,4 @@
 # limitations under the License.
 #
 
-sbt.version=1.3.4
+sbt.version=1.3.13
